Remove obsolete static_cast
The getGlowIntensity function now returns a value of the correct type.

+++ src/States/MapViewStateDraw.cpp
@@ -139,7 +139,7 @@ void MapViewState::drawResourceInfo()
 
 	const auto glowStep = getGlowIntensity();
 
-	const auto glowColor = NAS2D::Color{255, static_cast<uint8_t>(glowStep), static_cast<uint8_t>(glowStep)};
+	const auto glowColor = NAS2D::Color{255, glowStep, glowStep};
 
 	// Common Metals
 	renderer.drawSubImage(mUiIcons, position.x(), position.y(), 64, 16, constants::RESOURCE_ICON_SIZE, constants::RESOURCE_ICON_SIZE);
